services/voice/web_agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.9.0
zhipuai>=2.0.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.19.0

